only respond to latest messages on startup initial integration with OpenAI api

+++ JonasBot/Telegram/TelegramApi.cs
@@ -29,7 +29,7 @@ public class TelegramApi
         return new BotInfo(content.result.first_name, content.result.is_bot);
     }
 
-    public async Task<GetUpdatesDto> GetUpdates(int offset = 0)
+    public async Task<GetUpdatesDto> GetUpdates(long offset = 0)
     {
         var response = await client.GetAsync(GetUrl(ApiMethod.getUpdates) + $"?offset={offset + 1}");
         response.EnsureSuccessStatusCode();
